From: Henrik Nordstrom <henrik.nordstrom@ida.his.se>
Here is a patch to ftpget.c.

Effect: Negative or missing banner message from the ftp server is regarded
as a connection failure, and not login failure.

Reason(s):
1. ftpget didn't count the number of retries when the banner failed, resulting
in a endless loop if the server refuses to serve the connection. (accepts
the connection, but immediately closes the connection or sends a error
message)
2. The case where the connection is immediately closed, is really a connect
failure by a tcp wrapper, not a login failure.

The loop caused about 3000 emails to be sent from our wrapper. Our admin
got a bit frustrated about this.