This patch to alpha9 allows for large requests by the clients (for example POST file upload)
It adds the configuration directive request_size, which sets the
maximum request size.

The memory used for the request is grown in ASCII_REQUEST_BLOCKSIZE
steps (8Kb).

This patch is not very tested, but it should work... please test.

There is also a small change in cache_cf.c, to remove a compiler warning
when the in_addr localhost variable is initialized. (uninitialized members
in a initialized structure is set to binary 0 by standard)

/Henrik