Together with my earlier patch for correct parsing of ftp URLs, this patch fixes the problem with spaces in ftp URLs.
How:
The earlier patch casues path (and login info) to be in escaped form.
This makes sure that a space inside the path (or login info) is not
misinterpreated as a argument separator in the string sent to the
ftpget server.

This patch adds the decoding of escaped arguments to the ftpserver,
making the handling of ftp URLs with space complete.

/Henrik