The limit set for small allocations (M_MXFAST) is a bit to high.
It is currently set on 4096 bytes, and 100 blocks in each chunk. This
gives a very large space waste in the larger sizes.

A more reasonable limit is 1024. Anything larger than 1024 will result
in huge process sizes where only a fraction (around 50%) of the allocated
process space is used.

If one allocation is made of each size between 16 and 4096 bytes
(size in step of 16 bytes), 522240 bytes are allocated but the
process has grown with 52Mb. The additional size is because
malloc allocates memory in chunks of 100 blocks of each size.

This patch changes M_MXSIZE to 1024, and if compiled with -DXMALLOC_STATISTICS
some memory allocation size statistics are shown at the infopage.

Another thing is that memory not allocated with xmalloc should not be
freed with xfree since this breaks the XMALLOC_DEBUG (this was done in
ipcache.c where the result from tmpname was freed with xfree).

A short run with this statistics enabled shows that almost all allocations
are <1024, and the rest >=4096. There are very few allocations in the
range 1024-4096.

I'll run our server with this statistics enabled, and return with some
more interesting figures tomorrow.

/Henrik