Here is a small patch that enables one to use Electric Fence on squid.
Electric Fence is a simple malloc debugger, that sets up protected zones
after (or before) allocated memory, and invalidates freed memory.

If anyone is interested, I have modifications to Electric Fence for running
on Solaris2.x (even multithreaded applications), and buildling efence as
a shared library. Having efence as a shared library enables one to test
any application without recompiling, by using LD_PRELOAD (this is ofcourse
not possible if another version of malloc() is statically linked into the
application).

This patch extends the option -C to not close all unknown fildescriptors,
by the assuption that anyone that do not want to catch signals, is probably
debugging squid...

/Henrik