From:    Neil Murray <neil@aone.com.au> I was doing some testing and found a case I hadn't allowed for in my cachemgr.c mods.
Where you have a "?<proxy host>" without the ":" the code will play
up.  The change is below and in case it is easier for you the entire
patch is below that.  Assuming a virgin cachemgr.c from alpha 18 or
beta 1.

This also gives you the behaviour if that no ":<port>" is part of the
QUERY_STRING then it defaults back to the compiled port number, which
is what I think most users would expect.  I could have made it so that
a "?:" defaults back to the compiled proxy hostname but decided that
this way "?:" will wipe any compiled hostname defaults and you are
presented with a blank set of dialog boxes.