From:    John Saunders <John.Saunders@scitec.com.au> I just installed 1.0.13 and noticed that the shadow password support has stopped working. I took a look at the code and noticed the changes. The main problem is HAVE_LIB_SHADOW should be HAVE_LIBSHADOW.
I also made some changes that switch to suid mode when reading the normal
password file. This allows shadow password support on systems like NetBSD
and FreeBSD (and a few others that I don't know about) that transparently
read the shadowed password if uid == 0 when getpwnam() is called. I've
tested it on NetBSD and it works fine.