From:    Ron Gomes <rrg@ny.ubs.com>
We had problems with 1.1.9 out of the box.  They boiled down to this:
unlinkd is apparently required for proper operation, but there's no way
to configure a pathname other than /usr/local/squid/bin/unlinkd, nor
does squid notice or complain if unlinkd fails to start.

The code that follows (in the next message) makes these fixes:

- Adds "unlinkd_program" and "pinger_program" configuration options.
  (I noticed there's also no way to configure an alternate pathname for
  pinger, so I made that change as well.)  You may want to provide more
  elaborate comments than I did for the conf file.

- Makes sure that, if the exec of unlinkd fails, that fact is logged.
  (Previously the cache.log file descriptor was closed, so any such log
  entry went into the bit bucket.)

- Sets up an initial handshake (using another pipe) between squid and
  unlinkd so that squid can be sure unlinkd actually started, and shut
  itself down if not.