From:    David Luyer <luyer@ucs.uwa.edu.au>
Unless you completely and utterly trust shadowed passwords, SecureWare,
NIS, or whatever you use (all of which I've found holes with in
installations I've seen or commercial unix variants I've used), this is a
_very_ bad thing.

A user who obtains the _crypt_ of another user can, without having to
brute-force crack it, just pass (username, crypt) as a request for proxy
authentification.  (I have tested this and it works.  Not good.)

My fix is to add 128 to the first byte in the plaintext password which
squid stores to speed checks (on 2nd and subsequent logins).