From:    Henrik Nordstrom <hno@hem.passagen.se> Here is a patch that fixes FTP directory listings on some FTP servers / directories.
* Always include .. first in directory listings, regardless if the
origin server lists one or not. This fixes both servers that don't list
.., and directories that have strange files/directories that get listed
before ..  (UNIX example: ".-test")

* Show a "empty" listing when the origin server returns a empty listing.
This is much more intiutive than returning a 0-bytes object (which gives
a "document contains no data" error box in most browsers). This is for
example seen on protected directories where the user is not allowed to
list the files (like bin and etc in most anon-ftp setups).