we have FMR bugs with peer *'s in ps_state structure.   When a reconfigure occurs during ICP queries, and we have a timeout, the ->first_parent_miss peer will have been freed.
Using cbdata here would be too ugly.  we would have a lot of locks and
unlocks, plus what to do when the first_parent_miss peer is not valid?
re-select?

This approach saves sockaddr_in values for the peers.  We look up the
actual peer structure with whichPeer() when we really need the peers.