HEAD fixes (damn I hate HEAD)
on the server-side, HEAD transfers are always done soon as we make it
to the end of the reply headers

on the client-side, we can't do IMS for HEAD.  Always return cache HITS,
never a refresh.

httpMaybeRemovePublic() gets bastardized already (thanks henrik).  Also
make it remove HEAD objects for new GET requests.