Changed httpParseRequest to only look at the request headers. fixes the new problem where PUT/POST request was rejected that was intruduced by the NULL guard below.
Barf if request line/headers contains a NULL character.
C can't handle NULL characters very well, and it is not legal
to have control characters (including NULL) in the headers.
This fixes a assertion failure at another location in the code
on request with NULL characters in the headers.

Only look for \n as line terminator. \r should be ignored.
This is according to the recommendation in the HTTP
specification, and it also simplifies things. Fixes the
assertion failure on "simple" request without request headers
and \n as line terminator.