This patch fully transforms proxy_auth into a ACL type, allowing both allow and deny in any order suitable.
Authentication is requested from the user if:
* proxy_auth is used and no valid authentication header is present.
* wrong password is used
* denied by a proxy_auth ACL

A positive sideeffect is that is is also possible to use deny_info on
proxy_auth ACLs to change the message returned. It also fixes the
problem where proxy_auth user wasn't logged if the proxy_auth user info
wasn't cached.

This patch fully replaces my previous attemt.

There is a small collision with one of my other changes in proto.h, but
I assume that you know how to resolve this.