From:    Henrik Nordstrom <hno@hem.passagen.se>
* Changed async_io.c and aiops.c to use MemPool pools for their
"request" structures.

* Fix for a odd situation where I/O operations could get stuck in the
queue if there was no new I/O activity. Symptoms seen was: 1. The
shutdown code could trigger some assertions which seemed impossible. 2.
Some times a number of files could be kept open until the next swapout.

* Changed the log message shown when running low on I/O threads, and
also changed logging of this message to only occur if out of threads for
more than one second.

* Moved some aiops.c log messages to lower priorities.

* Removed that special low-priority queue for unlinks. I had some
strange effects from it, and the number of queued unlinks fluctuaded a
bit to much when under heavy load. If store maintaince is a problem then
this should it be corrected at a higher level, and not by delaying the
unlinks.

* Minor preparations for making NUMTHREADS a squid.conf tunable.

* moved memClean later in the shutdown procedure, to allow async-io to
use MemPool..