From:    Alfredo Sola <alfredo@intelideas.com>
If you install Squid listening in a secondary IP address instead
of the primary of the interface, it will generate for FTP listings
pages that refer icons to the primary. Given that Squid is not
listening on that interface, it results in a long timeout before
the page displays and "broken graphics" My suggestion is that
visible_hostname default to whatever tcp_incoming_address is,
instead of defaulting to the canonical name of the machine.

From:    Henrik Nordstrom <hno@hem.passagen.se>

<a patch that does a reverse lookup (BLOCKING!!) if tcp_incoming_address
is set.  The answer is used as cache hostname if the lookup succeeds.>