Two storeSwapOut() changes:
1) with new store-FS scheme, we don't get write callbacks.  This means
we were getting some objects "stuck" because we had, say 32K in memory,
but only wrote 8K in storeSwapOut() and there was no signal to make
us write another block.  Now we write as many blocks as we can,
instead of just one.

2) Re-fixed the case where storeSwapOut() gets called because the
store_status has changed to STORE_OK, but there is no data to swap
out.  Now close the swapout file, as done at the end of storeSwapOut().