aclHostDomainCompare had some rather nasty bugs.
We started the character-by-character comparison at the
wrong place.  First we compared the terminating NULL
bytes!  Further, this off-by-one bug caused us to stop
too soon and domains that differ in the first character
were called the same.

Also the check for '.' as the first character used OR logic
instead of AND logic.  So if the DOMAIN was shorter than
the HOST, and equal after the first character, it always
returned "greater than."