From: Damien Miller <damien@ibs.com.au> I just noticed that ARP acls were failing with squid-2.2-STABLE4 running on Linux 2.2.12. A bit of investigation revealed why:
It seems that 2.2.x uses a per-interface arp/neighbour cache, whereas
2.0.x uses a unified cache. Under 2.2.x you are required to specify a
interface name when looking up ARP table entries with SIOCGARP.
Squid's ARP acl code did not do this.

Attached is a patch which Works For Me. It retains the existing
behaviour under 2.0.x kernels by doing the ARP lookup with a blank
interface name.

If this fails, it will then try the ARP lookup for each interface that
the kernel knows about (up to a max of 32) excluding loopback and
alias interfaces.