Removing the "state == NULL" assertion in comm_write().
Previously, this assertion was always true because we used
safe_free(state), but then state got MemPool-ified.  Although
we did NULL fd_table[fd].rwstate, we didn't NULL 'state'.

The assertion is unneeded because fd_table[fd].rwstate will
always be freed if it is set at the beginning of the function.