Bugzilla #463: GET requests with request entity fail
This patch adds a new squid.conf directive "request_entities on/off".
If set to "on" then Squid will allow GET/HEAD requests with request
entities, even if such entites are "undefined" in the HTTP specification.

Technically speaking Squid is violating the HTTP specification when
not allowing such requests. The official way to handle such requests
is that the server should ignore the request entity.