Bug #648: deny_info TCP_RESET not working
This patch is the deny_info_url patch which corrects this issue and also
adds the ability to redirect. The earlier merge of the TCP_RESET deny_info
syntax from deny_info_url was not complete and did not work.

It was not originally planned to add the redirect capability to Squid-2.5,
but the patch is well tested and making a new patch which only fixes
TCP_RESET is not worth the effort.