Bug #856: Icon URLs are uneededly complex
The use of absolute URLs for the icons was not very good. While this
tries to limit the risk of icons failing in complex hetrogenous cache
hierarchies it is a common cause of problems

 - Clients often don't know how to address the proxy by hostname
 - When the clients do know, authentication gets messed up due to clients
   often regarding the proxy as a local server for which no proxy is required

This patch adds a "short_icon_urls" option which can be used to enable
the use of a shorter URL format (just the URL-path, not the hostname/port)
which avoids both issues.

However, there is a bug in older Squid releases where it does not detect
such icon requests correctly so this can not be set as the default..