Bug #922: random assertion failure if using authentication in miss_access, delay_access and a number of other directives, and a number of other fixes.
This patch is a second major cleanup of the authenticaiton framework
to make authentication persistent for the request. The major change
introduced here is that request->auth_user_request is the primary
auth_user_request structure, and checklist->auth_user_request is only
used internally by the auth code (and some safeguards in acl async lookups)

This should hopefully also get rid of some of the memory leaks seen earlier.