Bug #910: auth_user_hash_pointer "leak" in ntlm auth
This patch disables the response cache when challenge reuse is not
enabled, thereby avoiding the auth_user_hash_pointer usage.

There will still be a temporary leak of auth_user_hash_pointer if
challenge reuses are enabled in the ntlm configuration. The memory
is automatically reclaimed when the user expires from the auth
cache (authenticate_ttl)