Bug #910: Temporary memory leak in NTLM authentication when challenge is updated
This is a second attempt to apply this patch to Squid-2.5, this time after
fixing some bugs found the first time.

This patch makes sure to free the NTLM response cache entries when their
corresponding challenge is invalidated. Before this the NTLM response cache
grew in size for as long as the user was known to Squid (authenticate_ttl).