Bug #1166: The meaning of the configuration gets very confusing when we encounter empty ACLs such as
  acl something src "/path/to/empty_file.txt"
  http_access allow something somewhere

gets parsed (with warnings) as

  http_access allow somwhere

And similarily if you are using proxy_auth acls without having any auth schemes
defined.

This patch changes the parser to

  1. Allow empty acls to be defined and use them as such (no matching members).

  2. Using proxy_auth or related ACLs without auth schemes defined should be a
fatal errors. It is nontrivial for us to guess what the administrator intended
in such case, and most likely is an configuration error anyway.