Bug #1177: Don't close all "other" filedescriptors
Previously, when Squid was started it forcibly closed all "other" filedescriptors
other than stdin/stdout/stderr. While this is a reasonable security precaution
to clean up filedescriptor leakage from the caller it crashes some SSL libraries
and possibly other functions which opens internal filedescriptors on startup or
while the configuration is parsed (syslog likely candidate)

The reasoning in removing this function from Squid is that if the one starting
Squid has other filedescriptors open and not closing them this is their problem,
not ours.