Bug #1269: CONNECT requests truncated if client side disconnects first
This mainly causes problems for applications abusing the CONNECT method
for tunneling other traffic than SSL via the proxy, for example some FTP
clients when uploading files.

The "problem" was introduced by squid-2.5.STABLE6-CONNECT.patch which
immediately disconnects from the server when seeing a client disconnect
not waiting for pending "upload" data to be sent first.

It is strongly recommended to not use the CONNECT method in this manner.
If you want a general purpose proxy then look into SOCKS which provides
much better support for this kind of proxying.

Or in the case of FTP use a FTP proxy.