Bug #1223: Make the use of the %m error page to return auth info messages
This patch extends the helper protocols for Basic and Digest to provide
some basig information in error responses, and makes use of the error
response already included in the NTLM helper protocol, making these
messages available as %m in error pages. Can be used if desired to
indicate why a login failed. The exact messages returned is helper
dependent.