From: Nicolas Ruiz <nicolas@ula.ve>
This is a patch against squid-2.5.STABLE10-20050603 that does the
following:

- Translates errors/Spanish/ERR_INVALID_RESP to Spanish
- Replaces some ISO8859-1 characters by the preferred HTML code (i.e.
  "�" for "&aacute;").
- Minor typos or style fixes.
- Unifies the gender of "cache".
- Unifies the spelling of "cache" in spanish.

The last two changes deserve an explanation. The word "cache" in the
sense of a "web cache" or "cache memory" has not been formally (i.e. by
the "Real Academia de la Lengua Espa�ola" (Royal Academy of the Spanish
Language)) incorporated into spanish, therefore there isn't a sanctioned
spelling of the word. Nevertheless, the word is widely used in spanish,
but with inconsistencies: it is found both as a masculine and femenine
noun (spanish has only two genders, there is no neuter (it) pronoun) and
with at least two spellings. After a brief survey of the web, I set on a
single gender and spelling and fixed those error messages that were
using different ones (by virtue of being translated by different
people).