ipcache statistics cleanups
remove references to 'IpcacheStats.ghbn_calls'.  This counter was
only incremented in the ipcache_testnames() function, and not in
any other way useful.  Since the MIB includes an OID for this value,
always report 0 for it now.

remove references to IpcacheStats.release_locked.  It was never
incremented in the current code.

Remove IpcacheStats.errors.  It is never referenced or incremented
anywhere.

Resurrected from old 2.6. By Duane.