Vary/ETag updates & workarounds
broken_vary_encoding directive, for servers with broken ETag:s on Vary
content-encoding responses. I.e. servers sending the same ETag on recoded
responses as on plain responses.

Refresh logics corrected to not require If-Modified-Since. As sideeffect
all expired content now results in a conditional request, using
Date as source for If-Modified-Since if nothing else is known.