Bug #1677: Duplicate etags in the If-None-Match
If the client sent a If-None-Match header to Squid then duplicate ETag
values may be seen.

This change also works around a bug in Lighttpd by sending only a single
If-None-Match header possibly with a list of ETag values instead of one
header per value. The two forms are semantically equal, but Lighttpd responds
with 400 Invalid Request on the second..