Bug #1795: Theoretical memory leak in storeSetPublicKey
a code error in storeSetPublicKey potentially could cause a memory leak
if other parts of the code was changed in how Squid deals with Vary objects.

Not a problem in current releases as the failing code path is never reached
under normal operations, but could become a problem in future or with certain
patches.