Bug #1792: max_user_ip not working with ntlm_auth (patch 2)
This second patch to Bug #1792 really fixes the problem. The first patch
while good and making the code more sane did not really address the problem,
only one symptom of it.

The first patch is also kept as it does improve the code and makes room
for other related work in the authentication area.