Fix kqueue handling the almost theoretical case of overflowing the changes queue
kevent() does not like being called with an changes list alone. If there is
a single error in the changes list (i.e. an closed fd or similar) it stops
immediately and it's impossible to know what got processed or not..

this patch reallocs the arrays as needed to fit the current workload,
leaving only a single kevent call once per comm loop.