MFC: Bug #1915: assertion failed: client_side.c:4055: "buf != NULL || !conn->body.request"
This patch backs out a change in how Squid processes PUT/POST requests
when the server finishes the response before the complete request has
been sent. This change slipped in as part of the support for HTTP/1.1
chunked encoding. Was not meant to be there.

The change as such is probably correct, but other parts of the code
does not deal well with this situation.

Merged changes:
2007/03/12 21:56:55 hno +5 -2   Bug #1915: assertion failed: client_side.c:4055: "buf != NULL || !conn->body.request"