Author: Mark Bergsma <mark@wikimedia.org>
COSS assertion failure t->length == MD5_DIGEST_CHARS

This patch changes modifies some of the COSS rebuild code to not assert on
certain corrupt objects, but instead log an error and ignore the object.
Crashes are especially unhelpful because Squid will restart and keep crashing
in a loop, until manual intervention in the form of a cache clear is done.