Fix header cache, affecting content-length in access.log
The header cache was only partially set on cloned replies, causing
the content-type to not be logged in access.log.

This change makes the header cache to be fully initialized when the
reply is cloned, not even attempting to copy the cached results.

Could be optimized to copy the cached results instead.. probably
only content_length missing (audit needed).