MFC: Temporary shortage of system filedescriptors may cause Squid to permanently stop accepting connections
To adjust to system limits Squid automatically reduces the amount of
filedescriptors when it detects a system wide shortage. However, this
reduction might go so far that Squid completely stops accepting new
connections.

This patch adds a limit causing Squid to restart if the limit gets way
too low, in an attempt to recover if the situation is temporary.

Merged changes:
2007/06/25 14:38:00	hno	+3 -1	Bug #1798: Temporary shortage of system filedescriptors may cause Squid to permanently stop accepting connections