Author: Dave Dykstra <dwd@fnal.gov> -I command line option to accept open socket on stdin
This is a patch to add a '-I' option to squid that overrides the configured
http port with a bound socket passed in on standard input, for starting Squid
from inetd or similar port monitor.

This also allows for a method having multiple Squid instances listening on
the same port.