Fully initialise mem_node's now that they may be allocated non-zero'ed.
mem_node.uses wasn't being initialised and this would cause them to not be
free()d. It also exposes a little interesting bug which shows the tracking
of "cache_mem" can get out of sync with the actual allocation of stmem's
as my squid showed a mem cache of 100k when 150meg of entries were in use.
This may deserve some looking into at a later date.