external_refresh_check helper, allowing program control over cache refreshes
this helper is called whenever Squid needs to refresh a cached entry,
and allows full control on if the refresh really needs to be performed
to the backend server or if the cached copy is still fresh.

If fresh new header details may also be returned by the helper, updating
the cached headers.