Bug #2175: Update valgrind support for valgrind-3.3.0
Valgrind memcheck rearranged the names of it's support macros in
version 3.2.0 to better represent the actual function, and now the
old names has been removed starting with version 3.3.0 causing
build a failure if --with-valgrind-debug is used.

this patch updates Squid to use the new valgrind macro names, and
adds a little glue to be compatible with older valgrind versions
(tested with 1.9.something).