Tidy up the asyncio code whilst removing MEM_LINK_LIST
The initial aim was to remove MEM_LINK_LIST. It turns out the only
user (I can find!) of the linklist code (which implements a fifo
queue) is the aufs storeIOState readop/writeop queue.

Reimplement that using the new dlinkRemoveHead() primitive which
just removes the head node and returns it, or NULL on empty list.

This saves a memory allocation per aufs fileop and removes the
MEM_LINK_LIST mempool type.