Strip out the MEM_DLINK_NODE type and reorganise the helper code to live with out.
The helper code creates dlink nodes to attach requests to. Since a request
can be enqueued on 0 or 1 queues its a waste of the allocator. Throw a dlink_node
into the request/stateful_request structs and use that to populate the list.

This completes part one of the src/mem.c shuffle.

TODO: test the helper code! really test it!