Modify the IPv4 ACL code to use sqaddr_t's instead of in_addr's; prepare for IPv6 ACLs.
This approach differs somewhat from the Squid-3 IPv6 work. In Squid-3, IPv4
addresses are represented internally as V4-in-V6 addresses; in Cacheboy, I'm
keeping a clear distinction between IPv4 and IPv6 addresses for now.

This separate distinction may later prove to be slightly horrible - for example,
if a bound socket hands over an IPv4 connection wrapped inside an IPv6 socket
type and transition address, rather than the current dual-stack setup - but
that is a later problem to solve and will be easier to solve once the internals
are all IPv6 aware (as the mechanism for handling IPv4 will have been
abstracted away inside libsqinet and the internal representation is suddenly
irrelevant.)

The current changes:

* change aclIpDataPool() to use sqaddr_t instead of struct in_addr;
* rename decode_addr() to decode_v4_addr(); so it only decodes v4 addresses;
* aclMatchIp() still only takes an in_addr; it matches only on IPv4 addresses
  for the time being;
* aclParseIpData() still only does IPv4 DNS lookups and IPv4 record generation,
  even though its mostly IPv4/IPv6 agnostic.

Todo:

* Write myip6/src6/dst6 ACL type, and shoehorn it into the current setup (this is one
  area where Squid-3 has the leg up; writing new ACL types is much easier..)

* Modify aclMatchIp() to take a sqaddr_t * instead of an in_addr; this requires
  a few changes:

  - src6: checklist->src_addr needs to become an sqaddr_t; and thus everything which
    uses it (including fqdncache_gethostbyaddr())
  - myip6: checklist->my_addr needs to become an sqaddr_t
  - dst6: requires the ipcache code to be ipv4/ipv6 aware; "dst" currently
    performs ipcache_gethostbyname() lookups on request_t->host and then compares
    each of the replies against the ACL