Implement a simple but hackish method for delaying forwarding of request body requests.
This patch implements a new configuration parameter which controls whether a request
with a request body is forwarded immediately or whether it is delayed until enough
of the request body has been read in.

It uses the connection buffer to hold the request body data instead of trying
to extend the request body pipeline code. I'd rather sit this feature in the request
body pipeline handling but the current implementation makes it difficult to do
cleanly. Using the connection read buffer for this isn't -that- bad ; I'm just
breaking my own concept of 'layering' in the code.

The real trickery is in clientReadDefer() - where the request body would never
get bigger than a few kilobytes by default. This work patches that to allow it
to grow until it fills the pre-arranged size if the request is delayed. The
defer stuff kicks in once the request buffer is big enough and the request
is forwarded (and thus 'undelayed'.)