Change HttpHeaderEntry functions to operate on an allocated HttpHeaderEntry *, rather than create and free entries.
This is in preparation for migrating the HttpHeaderEntry array in a HttpHeader
over to a local array rather than the way its currently done. There's going to be
some dirty stuff done with regards to deleting items in arrays and such, but
it will drastically reduce the number of allocator calls made and hopefully
keep HttpHeaderEntry's in contiguous RAM, making searching easier and faster.

I'm pretty certain I've put the relevant alloc/free calls in.