Initial import of the reference counted String backing hack.
* String now uses buf_t instead of a pool'ed string buffer
* buf_t handles the NUL termination and reference counting
* Strings are copy on write'able
* Init'ing a String from another should be done via strDup() (which is a refcount
  operation) versus something like stringLimitinit(strBuf(), strLen()) (which
  is a copy)
* There's now scope for _STATIC_ string initialisation and reference counting,
  which leads to interesting little optimizations for header creation!
* Add in include's for buf.h for src/squid.h and libhttp/*.h where appropriate
* .. and the test casec ode.
* change something in the StrList.c code to use strIsNull() rather than
  if (strBuf() ..)

The existing code in src/ and libhttp/ grabs a const char * pointer to a string
buffer. Hopefully there isn't any code which does that and then continues to
after the original String is changed/reset/etc. This code maintains that
assumption.

This increases the memory operations required for creating new strings
(a buf_t creation and a string creation) and currently does -not- use the
mempool'ed string API in libmem/ . This however makes Dup'ing strings
extremely cheap.