Further comment the data flow in the http.c code.
This enumerates how the HTTP/0.9 code path works. the size/old_size thing
is merely a hack; its not actually all that required as far as I can tell.
The entire reply_hdr membuf is appended as a reply body as much as the
data left in the socket buffer is.

Next, is figuring out how to tidy up the 100 "continue" reply code in that
routine so the recursion doesn't involve that hackery, and we'll be set
to move onto the next phase.