Introduce the new testing httpReadReply() code; fix some compile problems.
The new code is (hopefully!) better structured than the old reply code.
NOTE: It doesn't yet handle 1xx responses; that needs to go in.

The new code does its appending -outside- of the reply parsing routine.
This is a precursor to removing that buffer copy step which seems to
hamper small-object proxy performance on slower hardware w/ slow FSB'es.
(Since it copies the whole reply into the membuf, then trims out the
bit it needs for the headers.)

There's a few things to do (like test!) the new code path before its deployed;
but this is really the bulk of the change. The rest of the change involves
using a MemBuf socket read routine (to be written!) and then convert over
the routine to instead use the left-over data in there, rather than in that
incoming socket buffer.

TODO:

* test, heh
* verify HTTP/0.9 replies work
* verify HTTP/1.0 / HTTP/1.1 works ok, with / without errors
* implement the memBufRead() routine (thats next)
* extend stuff to allow a "start" point inside the MemBuf for
  parsing the reply (and then obviously to start the copy for the
  rest of the reply body) - this is to allow both the skipping for
  1xx responses and the "skip whitespace" logic.
* convert over the code to read directly into that buffer, and then
  call httpAppendyBody() on the left-over chunk.