From Cacheboy:
Fix a problematic issue where requests are being created with a NULL method.

In the past, this would've been METHOD_NONE, but Benno didn't keep that
equivalence
in the code.

This fix attempts to brush over the issue by treating NULL methods as
"METHOD_NONE" so code called further down the pipeline doesn't break.