Remove the mempools -pools- code, replace with just direct calls to malloc/free.
The main reason why mempools saves CPU (and why this commit increases CPU use!)
is that just -way- too many calls to the memory allocator/free routines happen.

There are two reasons to remove this code:

* removing it will show which areas of the code calls the allocator too often;
* removing it will make writing threaded code easier down the track - I don't
  really think its a great idea to implement a really high performance, scalable
  allocator given it is 2009 and others do a much better job of this!

So yes, in the short term, CPU time will increase. In the longer term, the real
solution to this problem is to -reduce- the number of memory allocator calls
which are made, rather than trying to cache the calls which are going on.