Convert the httpStataData reply membuf to a buf_t.
This involves a whole lot of dirty code in src/http.c which absolutely should be nuked from orbit and really tidied up.

It also drops performance slightly - the incoming buffer isn't recycled for each chunk of data; instead its freed
and a new buffer is allocated. The eventual aim is to just reference the reply buffer for all of the reply strings
and reply body data to be referenced this, thus all of those malloc()s go away. Alas...