strBuf() -> stringDupToC() and safe_free()
Its unfortunate, but in this particular bit of code the rfc1738* routines are called
which take a const char * as an argument.

The need to convert the strings to C strings in these (and other) instances will
suddenly go away once the rfc1738* routines are modified to take and return a String.