Rename the AS routines into the BGP namespace.
Convert them to use a seperate radix base (is this actually a separate
radix base for each instance, or a shared backend instance with a root node
being exposed? nfi.)

This code allows networks to be added, but not (yet) to be deleted.
The old code never deleted individual items, it simply deleted the whole
table and started again.Sigh.