Migrate stuff thats not specifically -filedescriptor- oriented out of fd.c and into comm.c.
Speifically, these are calls to:

* commClose() - which removes the events from the event notification method (poll, kqueue, etc)
  - specifically needed for devpoll!
* Migrate the delay pools stuff out, its 100% socket only stuff.

Now, this may break for those few instances where FD's are not being treated correctly
as -comm- sockets (eg, the stuff that goes on in libiapp/disk.c) so the next few
commits should be to introduce debugging to code to assert() where things are being
used wrongly, and then to change those incorrect uses. :)

In any case, the next couple of commits will most likely be to exorcise out the
last bits of the comm loop abuse in libiapp/disk.c. I'm not sure how though.

(For posterity - the comm loop is used for pipes, fifos and sockets (which work)
and for sync disk IO calls (which return immediately, obviously) primarily to
do the work during the next pass through the event loop. Breaking up the disk work
to happen during subsequent passes through the comm layer is fine, but this mess
of different FD API uses is not. Files should use file methods, not FD_WRITE_METHOD/
FD_READ_METHOD directly, for example. Sigh.)