Bring over the last remaining string-reference patch from previous work.
This implements zero-copy NUL terminated strings. It doesn't allow strings
to be created from a larger buffer (ie, substrings are still expensive)
BUT it allows for cheap strDup().

It doesn't really save on anything at the moment - the memcpy()'s of
strings are gone but there's no reduction in the calls to malloc().
(Half the malloc() calls now that strdup is cheap; but double the malloc()
calls to create the buf_t and then the buffer.)

This combined with the HTTP vector work should begin dropping the amount
of needless malloc calls quite substantially.