The configure.in script was checking for the existance of an ERR_ACCESS_DENIED page in the error page subdireectory before automatically adding it to the list of default error page languages to install.
This broke when building the error page list including the new-school template
generated error pages because they don't exist early enough for ./configure
to pick them up.

Instead, populate a text file with the default error pages.

This also has the side effect of not installing templates/ in the destination
squid directory.