Disable update_headers by default.
update_headers is buggy and scales poorly - it copies the -entire- object just to update the headers.
This doesn't work well with very large objects that have been revalidated.

Some other solution should be investigated later. The only reason I'm not tossing the store
update code in its entirety is in case some users wish to continue using it.