Start building a "generate log from storedir somehow" tool.
The idea here (and this obviously doesn't yet build!) is to have one tool
that generates a binary swaplog on stdout. This tool will eventually be able
to choose between the swapdir contents, or the swaplog, or be forced to
pick one.

(Sidenote - i'd eventually prefer that all of the logic in this tool which
is reading the dirs and the files be moved into a library and reused between
a variety of smaller tools too. I'll think about that later.)