This is a quick experimental hack to see how quickly a directory can be walked if the files aren't stat()'ed.
The answer is "fast", by the way.

The rough aim here is to eventually provide a two-phase method of rebuilding from
the raw directory. Pass 1 would setup the bitmaps so the storedir can at least
have files placed in it (even though the storedir -may- still be overly full;
it can be pruned later on.) Phase 2 would then be to immediately allow reads
of objects from the store as they're read in.