Replace the storedir rebuild code with a call to the helper.
* open and close the helper using ipcCreate() / ipcClose()
* let the helper choose which is the best way to rebuild the storedir
* create a simple producer/consumer buffer (which should really be in
  a library! grr adrian) to handle the incoming objects to be handled.

The rebuild path still uses a non-threaded/non-async buffered write to
write to the temporary swaplog - so performance is still not anywhere
near what is possible.

The "rebuilding percentage" stuff isn't yet done either. I'm going to
rely on the helper inserting PROGRESS swaplog messages in the stream
to display. This hasn't yet been coded.

The actual old code hasn't been taken out - its been #if 0'ed out for now.
I'll do a drive by removal in a moment.

There is no handling of the COMPLETED swaplog message so it shows up
as a bad swaplog entry.

The rebuild process MUST finish on a COMPLETED message instead of an EOF;
that way Lusca knows the entire contents of the storedir have been slurped
in as best possible. I'll fix that in a subsequent commit.