Enforce the correct free function for the aio and coss per-swapIOState private data allocations
I'm not quite sure how this worked without crashing but I'm sure memory was being
free'd back into the wrong memory pool. This commit creates a free function in
storeIOState which gets called when the sio is free'd. This calls the correct
FS module free function.

This should resolve issue 29.