Add the client db entries into a linked list and walk -that- when removing radix entries
The radix tree implementation may remove and shuffle around nodes during a
radix_remove() call. This means that keeping pointers to radix_node_t's around
after a call to radix_remove() is not allowed.

Instead, use a dlink_list as well and walk that when determining which nodes need
to be removed.

This adds a radix lookup for each node being removed. Not a big deal.