Comment what the dstfwdip ACL will be.
This raises a good question - how will the ACL be properly evaluated?
If the ACL is parsed and handled before the HIT becomes a MISS, all kinds
of potential issues creep up.

This must thus be used not in http_access, but in miss_access. I'll have
to note that somewhere.