First phase of the store client reworking!
This commit changes the lifecycle of the store client code somewhat.
The basic idea is to keep the store client around long enough to
satisfy any outstanding disk IO requests. This will then allow the
async io to use an intermediary copy.

This code doesn't quite yet fully work - it works in the normal case,
but aborted/failed disk IO seems to tickle the completion paths
to be called and unwound in very bizarre methods.

Known TODO items:

* storeClientUnregister() is called in places where disk IO failed
  via the store IO "close" path, initiated by the store disk IO
  layer (ie, "aufs") when the disk IO fails. That needs to be
  fixed somehow.