Issue #99 - begin sorting out the method_t assignment related leaks and overwriting.
Modify the assignment code so overwritten method_t *'s are logged.
There's the unfortunate situation where already set pointers are
overwritten - and if they're METHOD_OTHER pointers, the string
value may leak.

It'll eventually be modified to free the previous value before
assigning it the new value; but that will change existing behaviour!