Issue #99 fixes - fix method_t pointer quirkiness
This commit modifies the behaviour of requestCreate() and code which
relies on copying method_t pointers around using urlMethodAssign().

Since urlMethodAssign() now may copy a method, rather than "gifting"
it to the called function, the behaviour of callers needed some
tinkering. In this specific case, parseHttpRequest() now needs to
call urlMethodFree() on the created method after parsing it.

In the past it would be just gifted to the request_t created somewhere
else; but urlMethodAssign() will now cheerfully create a private
duplicate just for that request.

For correctness, src/htcp.c also now explicitly calls urlMethodFree().
All calls to urlMethodGet*() should be followed by an explicit
urlMethodFree() now and I'll hopefully do this in some subsequent
commits.

urlMethodAssign() will now:

* free the previous destination pointer if it's set and if it's a
  METHOD_OTHER pointer;
* call urlMethodDup() on the src pointer rather than blindly copying
  the value over.

This modifies two things:

* If the assignment would overwrite a method_t pointer that is pointing
  do a METHOD_OTHER method, it'll first free it, preventing memory leaks;
* If the assignment is from a METHOD_OTHER method, it creates a new method
  and assigns that rather than point to the existing one.

This fixes the specific crash in Issue #99 where a method_t pointer, pointing
to a METHOD_OTHER method, is free'd before the method_t pointer is used
in the final request logging bit.

The performance and memory allocation use in the standard method case is
the same. There'll be a very slight increase in memory use when handling
non-normal methods.