Issue #108 - modify the entry parser to parse and add entries at the same time
The HTTP header entry parser just creates entries which may or may not be
added to the HTTP header set.

This commit modifies the HTTP header entry parser to parse and create a header
entry, then add it to the HttpHeader list.

The parse checking now occurs inside the header entry parser rather than the
header parser.