Begin trying to sort out the horrible mess that is AccessLogEntry.
The motivation here is to have Init and Done functions which correctly
call sqinet_init() / sqinet_done(). This is proving to be slightly more
difficult than it should be because of how AccessLogEntry is used
(ie, as a dumping ground for random bits of logging related information
which is also somewhat used elsewhere.)

This first commit includes IPv6-savvy address assignment API additions
but it in particular tries to define some semblence of an "object
lifecycle." There's plenty of ugly direct frobbing of AccessLogEntry
internals by various places in the code - this will have to be teased
apart at some point in the near future.