Fix a memory leak in the errorState handling when connections are aborted.
The ErrorState entry may not have been valid when this function is called.
If it is valid, it is only used if the StoreEntry doesn't yet have any data
fed into it.

This patch ensures the ErrorState is freed, regardless of whether it's used
(and then implictly freed) or not.