Merge r13332 from /playpen/ipv6_acl
Shuffle around the decode_v6_addr function order; first try and
parse the IPv6 address and, if that fails, try to parse out a
masklen.

getnameinfo() -  when not told what address family to parse - was
happily parsing an integer as an IPv4 address. This isn't what I
want. Luckily a lone integer isn't a valid IPv6 address? :)