Migrate request_t->my_addr / my_port to IPv6 aware types.
Squid/Lusca sets my_addr to "NoAddr" when request_t is created.
This new code leaves it unassigned so any access to it will error.
It must be set somewhere (eg by parsing a client request.)

The forward.c transparent code path uses request_t->my_addr
as a destination address for outbound connections if a DNS lookup
on the client-provided URL fails. This codepath is currently
IPv4-only. Which is fine; the forward code is not going to be made
IPv6-aware in this pass.