Convert request_t->out_ip to sqaddr_t.
A few places where this is set (notably the getOutgoingAddr() function)
still only return IPv4 addresses, so an IPv4 address is only going to be
set. This will change in the future as more of the code is taught about
IPv6.

It defaults to IPv4 ANY_ADDR to preserve the previous behaviour.