Flip forward.c over to use the new commConnect() API.
This may be reverted later on if I decide it's the wrong way to go.

The big change here is that the FD is now created inside the src/comm2.c
routines and returned if successful. The forward code now doesn't create
FDs itself.

It also means that there aren't sockets hanging around which are waiting
for a DNS resolution pass to happen. That always annoyed me in Squid.

Things that work:

* Combination IPv4/IPv6 destination hosts - it simply tries each
  entry in order that it's received in the ipcache struct

Things that don't work:

* setting the outbound address - since this can be v4 or v6,
  both have to be passed into the new commConnect() API for now.
  I'll worry about tidying that up later.

* Transparency - v4 or v6 transparency won't work

* Allowing the admin to choose whether v4 or v6 hosts are tried
  first

* cache peers are still v4 only